python-ai/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow>=10.2.0
ultralytics>=8.1.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
torch>=2.1.0
torchvision>=0.16.0

